Fixes Review Order when used with Roots Theme
the Roots Theme http://www.rootstheme.com/ sets 'template_url' for get_bloginfo() to the empty string, which causes issues with jigoshop_get_template_file_url() in its current state.  This commit ensures that the same file which file_exists() checked for is returned, instead of relying on get_bloginfo() to produce the correct path.

Additionally, this commit exchanges the one-line if syntax for the colon if/endif syntax used elsewhere in jigoshop.