add all_items label to the product and order post type, better german translation to keep it inline with wordpress
- added all_items label to keep the naming and translation of jigoshop
admin more inline with the latest wordpress version.
- in the german translation the Produktkategorien were replaced with
Kategorien. because this is the new naming scheme that wordpress is
using for posts, pages, etc. Produkt-Schlagworte is no Schlagwörter.

maybe the original english labels could also be renamed to the new
wordpress scheme. but this would mean some work in the code and on
every translation file.