added add_action on cart after plugins_loaded.
Modifications have been made to both cart and shipping classes to update
the action of calculate totals and calculate shipping.

Also, a sanity check on an array in the calculable shipping class so that
the first time it's loaded to admin, it doesn't throw an error when the
array hasn't been initialized.