Updates to calculable shipping
Fix to review_order and jigoshop_actions to include the service rate
id within the dropdown value so that in jigoshop_actions we can set
the session with selected_rate_id.

Fix to script.js at the cart when user is selecting their method of
shipping for the cart calculations. The animated gif was not showing
because of the new change to the params.