Fixes all-states for taxes
This fix allows a user to select all-states now and it will apply the
particular tax class to all states of a given country.
Change made for that is to create a full array of every state in the
country instead of the old mechanism of *. The reason the full list of
states has to be created is because of multiple tax classes being able to
be applied to a product.