Fixes estimated taxes for
Fix includes is customer shipping outside of base. Basically taxes are
only applied if the customer is shipping within country, but to a
different state/province.

If a customer is shipping out to another country, taxes will not apply.

Also, if no country is set, assume that shipping will happen to the shop
base country and/or state, and calculate taxes for the shop base, and say
taxes are estimated until the shipping country is supplied.