Provides basic error message handling from shipping
The current code had 1 error message at the cart and that was when the
there was no shipping for a customer, they were notified.

With calculable shipping, I have found that there are times when shipping
cannot be calculated (eg. missing state, or missing postalcode/zipcode).
Without this error messaging framework, the message to the user is
completely misleading, as it states there are no shipping services
provided to your country.

Now the shipping service itself can return error messages that get
propagated to the cart and to checkout.