Fixes #391 unwanted exclusion ...
This fix takes care of the fact that the order page shows prices excluding
tax...This is correct, but the problem actually occurs when someone adds a
new entry to their order, the code used to take the product price (which
would have been including taxes) and uses that price with the new
additional order item.

The entries are all supposed to be showing price excluding tax, as the
example shows in the placeholder tag. Therefore, when adding a new item,
it is necessary to check to see what the settings are for tax. If prices
include tax, then we need to get the product price excluding tax, rather
than the regular product price.