Fixes #138 and #139
for the cart breaking, we need to validate the chosen method is calculable
shipping, as the regular shipping_methods do not have the function
has_error()

the checkout broke because the assumption was made that a calculable
shipping method would be chosen and only showed the options when a
calculable shipping method was set to chosen. If regular shipping methods
were included and they got chosen, the selection of calculable shipping
would not display.

added new static function on jigoshop_checkout.class to handle the
dropdown entries as the code had been previously duplicated on the
review_order.