fixes subtotal on compounded taxes
When looking at the way the taxes are displayed for compounding taxes, I
noticed that subtotals do not include the taxes, but rather the prices
without tax. Therefore, by moving the non compounded tax below the
subtotal, I also needed to adjust the subtotal.

This fixes that, as well as cleans up the code. Ahh, so much nicer now.
All the extra work to put the tax ahead of subtotal now makes it simple
when it comes after the subtotal.