@mindbuster, revisits 29bfa6cab45906e10da45b385114cfedd931c5b4 .
is_in_stock() always returned true when manage_stock is 'no', so that last if statement wasn't even being ran since it's checking for a 'false' value.

Previous if block:

Is manage stock enabled?
 Yes: Is it not in stock or does it not have enough stock?
 No: Is it not in stock? -- It will never be in stock because manage stock is disabled xD

Current if block:

Is manage stock enabled? And is a product out of stock or not have enough stock?