Add Table Prefix on Jigoshop System Info
Allows us to check to see if pre-existing table names are beyond 
character limit in debugging downloadable and other types of products.
We should also check for length of prefix on both install and also on
upgrade. On upgrade, users should be informed that their prefix is too
long and should be given a link with more info (that link should go to
the Jigoshop Minimum System requirements page, ideally a anchor tag 
where we explain the length of prefix cannot be > 16 characters and 
how to fix the issue (use Database Prefix Changer plugin).