Contributing.md mention in readme
This is mostly so that we have a build on travis after we fixed that
subdependency. Because the hook was off at the time of pull, our image
on readme will be failed, because a build with the dependency back in
was never tested.  Recommend turning the hook on (and leaving it there),
and then merging this in.