Update to sale shortcode for sales between dates
Unless I've completely misunderstood, the [sale_products] shortcode was broken for sales between dates. 

The 'sale_price_dates_from' and 'sale_price_dates_to' are saved as timestamps in the database, but compared with a regular date, plus the compare operator was wrong for 'sale_price_dates_to'.

I noticed the problem on a site I run, where the [sale_products] would keep showing products after the sale was over, but with the original price. This is how I changed the code there, and it seems to work fine.