Validate URL + more
Thanks!

1). I added the description to the URL field as you said

2). I'm validating the URL when saving to DB now with the
FILTER_VALIDATE_URL but I'm not showing any issues in the GUI. I could
not find where to do that. Any help here?

3). I remove the birthday field for now. The user will have to fill
that out to Skrill instead.

4). I added translation to the confirmation note.
translation for confirmation_note

5). Yes. That would be great. I started there but there are so many
different options that I'm not sure if I'm able to pull that of. I
don't know which are allowed and not allowed together. Look at p40 here
https://www.moneybookers.com/merchant/en/moneybookers_gateway_manual.pdf
 Do you have any other idea? That is not that hard for me to implement?