Remove the transient cache
These transient caching lines aren't present in the other widgets and cause a problem when people are using WPML. The problem is when switching between language this widget stays in the base language. Removing the transient cache lines resolves the problem. I suggest having them permanently removed as they don't seem to effect anything with this widget.